pyproject.toml
[project]
name = "rotten-bot-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "asyncpg==0.30.0",
    "fastapi-mail==1.5.0",
    "fastapi[standard]==0.117.1",
    "itsdangerous==2.2.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.37.0",
    "opentelemetry-sdk==1.37.0",
    "passlib==1.7.4",
    "pydantic-settings==2.11.0",
    "pyjwt==2.10.1",
    "pytest==8.4.2",
    "redis==6.4.0",
    "sqlmodel==0.0.25",
]
